+++ satella/coding/structures/__init__.py
@@ -3,7 +3,7 @@ from .dictionaries import DictObject, apply_dict_object, DictionaryView, TwoWayD
 from .hashable_objects import HashableWrapper
 from .immutable import Immutable, frozendict
 from .singleton import Singleton, SingletonWithRegardsTo
-from .structures import SetHeap, OmniHashableMixin, TimeBasedHeap, Heap
+from .structures import SetHeap, OmniHashableMixin, TimeBasedHeap, Heap, ReprableObject
 from .proxy import Proxy
 from .typednamedtuple import typednamedtuple
 from .ranking import Ranking
@@ -12,6 +12,7 @@ from .sorted_list import SortedList, SliceableDeque
 __all__ = [
     'KeyAwareDefaultDict',
     'Proxy',
+    'ReprableObject',
     'DirtyDict',
     'SortedList',
     'SliceableDeque',