+++ satella/coding/structures/dictionaries.py
@@ -6,8 +6,6 @@ from satella.coding.recast_exceptions import rethrow_as
 from satella.configuration.schema import Descriptor, descriptor_from_dict
 from satella.exceptions import ConfigurationValidationError
 
-__all__ = ['DictObject', 'apply_dict_object', 'DictionaryView', 'TwoWayDictionary', 'DirtyDict',
-           'KeyAwareDefaultDict']
 
 K, V, T = tp.TypeVar('K'), tp.TypeVar('V'), tp.TypeVar('T')