+++ satella/coding/structures/ranking.py
@@ -3,8 +3,6 @@ import typing as tp
 
 from .sorted_list import SortedList
 
-__all__ = ['Ranking']
-
 
 T = tp.TypeVar('T')