+++ satella/coding/structures/sorted_list.py
@@ -1,8 +1,6 @@
 import collections
 import typing as tp
 
-__all__ = ['SortedList', 'SliceableDeque']
-
 T = tp.TypeVar('T')