+++ satella/coding/concurrent/__init__.py
@@ -3,6 +3,7 @@ from .locked_dataset import LockedDataset
 from .monitor import MonitorList, Monitor, MonitorDict, RMonitor
 from .locked_structure import LockedStructure
 from .thread import TerminableThread, Condition
+from .atomic import AtomicNumber
 
 __all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread',
-           'MonitorDict', 'MonitorList', 'Condition', 'LockedStructure']
+           'MonitorDict', 'MonitorList', 'Condition', 'LockedStructure', 'AtomicNumber']