+++ satella/coding/sequences/__init__.py
@@ -1,9 +1,9 @@
 from .choose import choose
 from .iterators import infinite_counter, take_n, is_instance, skip_first, zip_shifted, stop_after, \
-    iter_dict_of_list, shift, other_sequence_no_longer_than, count, even, odd
-from .sequences import is_last, add_next, half_cartesian, group_quantity
+    iter_dict_of_list, shift, other_sequence_no_longer_than, count, even, odd, n_th
+from .sequences import is_last, add_next, half_cartesian, group_quantity, Multirun
 
 __all__ = ['choose', 'infinite_counter', 'take_n', 'is_instance', 'is_last', 'add_next',
            'half_cartesian', 'skip_first', 'zip_shifted', 'stop_after', 'group_quantity',
-           'iter_dict_of_list', 'shift', 'other_sequence_no_longer_than', 'count',
-           'even', 'odd']
+           'iter_dict_of_list', 'shift', 'other_sequence_no_longer_than', 'count', 'n_th',
+           'even', 'odd', 'Multirun']