+++ tests/test_coding/test_concurrent.py
@@ -2,12 +2,25 @@ import platform
 import time
 import unittest
 import copy
+import threading
+import time
 
 from satella.coding.concurrent import TerminableThread, CallableGroup, Condition, MonitorList, \
     LockedStructure, AtomicNumber
 
 
 class TestConcurrent(unittest.TestCase):
+
+    def test_atomic_number_wait(self):
+        an = AtomicNumber()
+
+        def process_atomic_number(an):
+            time.sleep(0.3)
+            an += 1
+
+        self.assertEqual(an, 0)
+        threading.Thread(target=process_atomic_number, args=(an, )).start()
+        an.wait()
 
     def test_atomic_number(self):
         a = AtomicNumber(4)