+++ satella/coding/concurrent/__init__.py
@@ -1,4 +1,4 @@
-from .callablegroup import CallableGroup
+from .callablegroup import CallableGroup, CallNoOftenThan
 from .locked_dataset import LockedDataset
 from .monitor import MonitorList, Monitor, MonitorDict, RMonitor
 from .locked_structure import LockedStructure
@@ -6,4 +6,5 @@ from .thread import TerminableThread, Condition
 from .atomic import AtomicNumber
 
 __all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread',
-           'MonitorDict', 'MonitorList', 'Condition', 'LockedStructure', 'AtomicNumber']
+           'MonitorDict', 'MonitorList', 'Condition', 'LockedStructure', 'AtomicNumber',
+           'CallNoOftenThan']