+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # v2.7.23
 
 * added custom `time_getter_callable` for `satella.time`
-
-
+* fixed Prometheus escapes