+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.7.24
+
+* `Proxy` objects will suceed isinstance check on the objects that they wrap