+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.7.24
 
 * `Proxy` objects will suceed isinstance check on the objects that they wrap
+    * as a side effect both `Proxy` and `LockedDataset` are no longer Generic
+