+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.7.25
+
+* added `wait` to `AtomicNumber`