+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.7.25
 
 * added `wait` to `AtomicNumber`
+* added `Condition` to the docs